requirements.txt
fastapi
uvicorn
streamlit
requests
faiss-cpu
langchain
langchain-community
langchain-groq
sentence-transformers
numpy
PyYAML
mlflow
